tado_guestmode/requirements.txt
falcon==3.0.1
requests==2.27.1
pyyaml==6.0
gunicorn==20.1.0